requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9

# LangGraph and LangChain
langgraph>=0.0.20
langchain>=0.0.300
langchain-openai>=0.0.2
langchain-community>=0.0.10
langsmith>=0.0.60

# Testing
pytest>=7.4.2
httpx>=0.25.0

# Utilities
pydantic>=2.4.2
python-dotenv>=1.0.0
graphviz>=0.20.1

# Document processing
python-docx>=0.8.11
docx2txt>=0.8